\name{Lassosum.extension-package}
\alias{Lassosum.extension-package}
\alias{Lassosum.extension}
\docType{package}
\title{\packageTitle{Lassosum.extension}}
\description{\packageDescription{Lassosum.extension}}
\section{Package Content}{\packageIndices{Lassosum.extension}}
\author{\packageAuthor{Lassosum.extension}}
\section{Maintainer}{\packageMaintainer{Lassosum.extension}}
\keyword{package}
