\name{LassosumExtension-package}
\alias{LassosumExtension-package}
\alias{LassosumExtension}
\docType{package}
\title{\packageTitle{LassosumExtension}}
\description{\packageDescription{LassosumExtension}}
\section{Package Content}{\packageIndices{LassosumExtension}}
\author{\packageAuthor{LassosumExtension}}
\section{Maintainer}{\packageMaintainer{LassosumExtension}}
\keyword{package}
