\name{multivariateLassosum-package}
\alias{multivariateLassosum-package}
\alias{multivariateLassosum}
\docType{package}
\title{\packageTitle{multivariateLassosum}}
\description{\packageDescription{multivariateLassosum}}
\section{Package Content}{\packageIndices{multivariateLassosum}}
\author{\packageAuthor{multivariateLassosum}}
\section{Maintainer}{\packageMaintainer{multivariateLassosum}}
\keyword{package}
